work/nanopro/benchmarks/scikillbench/deep_solver_benchmark/requirements.txt
# CASCADE Requirements
# Python 3.12+ required
# For uv users: uv pip install -r requirements.txt

# ============== CORE DEPENDENCIES ==============

# --- AI/ML Core Frameworks ---
openai==1.102.0
openai-agents==0.2.8
transformers==4.56.0
huggingface-hub==0.34.4
tokenizers==0.22.0
tiktoken==0.11.0

# --- Materials Science/Chemistry Core ---
pymatgen==2025.6.14
pymatgen-db==2023.7.18 
ase==3.26.0
matminer==0.9.3
mp-api==0.45.4
mpcontribs-client==5.10.3
rdkit==2025.3.5
chgnet @ git+https://github.com/CederGroupHub/chgnet 
mace-torch==0.3.14
openbabel-wheel==3.1.1.22
robocrys @ git+https://github.com/hackingmaterials/robocrystallographer.git@4ee318f995ca249a29a89241f098014102d855ae
mlip==0.1.3
smact==3.2.0
emmet-core==0.84.9
spglib==2.6.0
sympy==1.14.0
mdf-forge==0.8.0
mdf-toolbox==0.7.1
obelix-data==1.2.0

# --- Deep Learning & Graph Neural Networks ---
torch==2.7.1
torch-geometric==2.6.1
torchaudio==2.7.1
torchdata==0.11.0
torchvision==0.22.1
lmdb==1.7.3

# --- Scientific Computing ---
numpy==1.26.4
scipy==1.16.1
pandas==2.2.3
scikit-learn==1.7.1
scikit-image==0.25.2
joblib==1.5.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
h5py==3.14.0
networkx==3.5
cycler==0.12.1
kiwisolver==1.4.9
fonttools==4.59.1

# --- Web Framework & API ---
fastapi==0.116.1
uvicorn==0.35.0
gunicorn==23.0.0
starlette==0.47.3
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1

# --- Database Clients ---
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
supabase==2.18.1
supabase-auth==2.12.3
neo4j==5.28.2
pymongo==4.10.1

# --- Multi-turn Chat & Memory ---
streamlit==1.43.2
mem0ai[graph]>=1.0.0  # Includes graph memory support (langchain-neo4j, kuzu, etc.)
vecs==0.4.5  # Vector database client for Supabase (required by Mem0)

# --- HTTP & Async ---
httpx==0.28.1
httpx-sse==0.4.1
aiohttp==3.12.15
aiofiles==24.1.0
aioitertools==0.12.0
aiosignal==1.4.0
requests==2.32.5
asyncio==4.0.0
anyio==4.10.0
nest-asyncio==1.6.0
certifi==2025.8.3
urllib3==2.5.0
charset-normalizer==3.4.3
idna==3.10
multidict==6.6.4
yarl==1.20.1
frozenlist==1.7.0

# --- MCP Protocol ---
mcp==1.13.0

# --- Search & Web Scraping ---
tavily-python==0.7.11  # For Tavily search API
crawl4ai==0.7.3
beautifulsoup4==4.13.4
selenium==4.35.0
playwright==1.54.0
lxml==5.4.0

# --- Configuration & Utilities ---
python-dotenv==1.1.1
pyyaml==6.0.2
omegaconf==2.3.0
click==8.2.1
rich==14.1.0
tqdm==4.67.1
psutil==7.0.0

# --- ML Experiment Tracking ---
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2

# --- Data Serialization ---
jsonschema==4.25.1
jsonlines==4.0.0
orjson==3.11.2
msgpack==1.1.1

# --- Security ---
cryptography==45.0.6
bcrypt==4.3.0
pyjwt==2.10.1

# --- Development Tools ---
ipython==9.4.0
setuptools==79.0.1
pip==25.0.1

# --- Additional Core Dependencies ---
monty==2025.3.3
packaging==24.2
protobuf==5.29.5
pillow==11.3.0
imageio==2.37.0
numba==0.61.2
jinja2==3.1.6
markupsafe==3.0.2
typing-extensions==4.14.0
attrs==25.3.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cython==3.1.3
decorator==5.2.1
dill==0.4.0
einops==0.8.1
executing==2.2.0
filelock==3.19.1
fsspec>=2024.9.0,<2025.0.0  # Constrained by neo4j-graphrag
cffi==1.17.1
chardet==5.2.0
blinker==1.9.0
cachetools==5.5.2
distro==1.9.0

# --- JAX Ecosystem (required for mlip) ---
jax==0.4.33
jaxlib==0.4.33
flax==0.9.0
optax==0.2.5
orbax==0.1.9
orbax-checkpoint==0.5.2
jraph==0.0.6.dev0
e3nn-jax==0.20.7
ml-dtypes==0.5.3
opt-einsum==3.4.0

